separate out save and load data
